requirements.txt
fastapi [all]
sqlalchemy
psycopg2
uvicorn
alembic
secure-smtplib
python-dotenv
loguru
fastapi_permissions
PyJWT
bcrypt
passlib
python-jose
ppillow
python-multipart